Removed fixed schema generation and merged all resources in one function
