+++ source-hubspot-native/source_hubspot_native/__init__.py
@@ -13,7 +13,6 @@ from estuary_cdk.capture import (
 from estuary_cdk.http import HTTPMixin
 
 from .resources import all_resources
-from .custom_resources import all_custom_resources
 from .models import (
     ConnectorState,
     EndpointConfig,
@@ -42,9 +41,6 @@ class Connector(
         self, log: Logger, discover: request.Discover[EndpointConfig]
     ) -> response.Discovered[ResourceConfig]:
         resources = await all_resources(log, self, discover.config)
-        custom_resources = await all_custom_resources(log, self, discover.config)
-        if custom_resources is not None:
-            resources += custom_resources
         return common.discovered(resources)
 
     async def validate(
